refactor: Voucher 프로그램과 관련된 enum 패키지를 생성 & 분리
