+++ src/main/java/com/prgrms/springbootbasic/enums/voucher/VoucherDeleteMenu.java
@@ -1,4 +1,4 @@
-package com.prgrms.springbootbasic.enums;
+package com.prgrms.springbootbasic.enums.voucher;
 
 public enum VoucherDeleteMenu {
     ID,
@@ -8,7 +8,7 @@ public enum VoucherDeleteMenu {
         try {
             return valueOf(deleteMenu.toUpperCase());
         } catch (IllegalArgumentException e) {
-            throw new IllegalArgumentException("잘못된 명령어를 입력하셨습니다! 실행하고자 하는 바우처 명령어를 다시 확인 해주세요!");
+            throw new IllegalArgumentException("잘못된 명령어를 입력하셨습니다! 삭제하고자 하는 바우처 프로그램의 명령어를 다시 확인 해주세요!");
         }
     }
 }
